PostgreSQLAdapter::Utils seems to be only used from a single spot - quite redundant
